store: Move advisory locks for migrations to separate module
